Split preview.rs into pure helpers and typer state machine
Separate the pure text-diff algorithms from the stateful, keyboard-driving
typer. preview.rs keeps the free-function helpers (preprocess/extension/
prefix/tail-match); the new typer.rs owns State + Typer.

Moving the keyboard-free state transitions onto impl State makes them
testable without a keyboard Simulator (previously impossible), adding 8
characterization tests that pin the formerly-untested state machine.
Behavior is unchanged.
